refactor : advice 적용
