+++ src/main/java/Dcoding/Celebrem/common/config/AppConfig.java
@@ -1,4 +1,4 @@
-package Dcoding.Celebrem.config;
+package Dcoding.Celebrem.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
