+++ src/main/java/Dcoding/Celebrem/common/config/EmailConfig.java
@@ -1,4 +1,4 @@
-package Dcoding.Celebrem.config;
+package Dcoding.Celebrem.common.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
