+++ src/main/java/Dcoding/Celebrem/common/exception/BadRequestException.java
@@ -1,4 +1,7 @@
 package Dcoding.Celebrem.common.exception;
 
-public class BadRequestException {
+public class BadRequestException extends RuntimeException{
+    public BadRequestException(final String message) {
+        super(message);
+    }
 }
