+++ src/main/java/Dcoding/Celebrem/common/exception/NotFoundException.java
@@ -1,2 +1,8 @@
-package Dcoding.Celebrem.common.exception;public class NotFoundException {
+package Dcoding.Celebrem.common.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(final String message) {
+        super(message);
+    }
 }
