+++ src/main/java/Dcoding/Celebrem/common/exception/UnauthorizedException.java
@@ -1,2 +1,8 @@
-package Dcoding.Celebrem.common.exception;public class UnauthorizedException {
+package Dcoding.Celebrem.common.exception;
+
+public class UnauthorizedException extends RuntimeException {
+
+    public UnauthorizedException(final String message) {
+        super(message);
+    }
 }
